ScreensPackage erstellt und Verwendung von Klasse Dice
