+++ app/src/main/java/com/example/mankomania/screens/CreateNewLobby.java
@@ -1,11 +1,9 @@
-package com.example.mankomania;
+package com.example.mankomania.screens;
 
 
 import android.content.Intent;
 import android.content.SharedPreferences;
 import android.os.Bundle;
-import android.util.Log;
-import android.view.View;
 import android.widget.ArrayAdapter;
 import android.widget.Button;
 import android.widget.CompoundButton;
@@ -20,6 +18,7 @@ import androidx.core.graphics.Insets;
 import androidx.core.view.ViewCompat;
 import androidx.core.view.WindowInsetsCompat;
 
+import com.example.mankomania.R;
 import com.example.mankomania.api.Lobby;
 import com.example.mankomania.api.Status;
