+++ app/src/main/java/com/example/mankomania/screens/PlayerViewModel.java
@@ -1,6 +1,5 @@
-package com.example.mankomania;
+package com.example.mankomania.screens;
 
-import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
 
 public class PlayerViewModel extends ViewModel {
