+++ app/src/main/res/layout/activity_board.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Board">
+    tools:context=".screens.Board">
     <RelativeLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
