+++ app/src/main/res/layout/activity_choose_your_character.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ChooseYourCharacter">
+    tools:context=".screens.ChooseYourCharacter">
 
     <TextView
         android:id="@+id/ChooseYourCharacter_heading"
