+++ app/src/main/res/layout/activity_create_new_lobby.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".CreateNewLobby">
+    tools:context=".screens.CreateNewLobby">
 
     <LinearLayout
         android:id="@+id/firstRow"
