+++ app/src/main/res/layout/activity_event_roll_dice.xml
@@ -4,7 +4,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Event_RollDice">
+    tools:context=".screens.Event_RollDice">
 
     <RelativeLayout
         android:layout_width="match_parent"
