+++ app/src/main/res/layout/activity_finances_and_stocks.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".FinancesAndStocks">
+    tools:context=".screens.FinancesAndStocks">
 
     <RelativeLayout
         android:id="@+id/relativeLayout"
@@ -31,7 +31,7 @@
 
                 <TextView
                     android:id="@+id/Finances_header"
-                    android:layout_width="wrap_content"
+                    android:layout_width="0dp"
                     android:layout_height="45sp"
                     android:layout_weight="1"
                     android:gravity="center"
