+++ app/src/main/res/layout/activity_game_score.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".GameScore">
+    tools:context=".screens.GameScore">
 
     <TextView
         android:id="@+id/GameScore_heading"
