+++ app/src/main/res/layout/activity_main_login.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainActivityLogin">
+    tools:context=".screens.MainActivityLogin">
 
   <Button
       android:id="@+id/Login_LoginButton"
