+++ app/src/test/java/com/example/mankomania/TestEmail.java
@@ -3,6 +3,8 @@ package com.example.mankomania;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
+import com.example.mankomania.screens.MainActivityLogin;
+
 import org.junit.jupiter.api.Test;
